Add an l3build script for graphics
This is essentially working but at the moment there are a few notes (see below).
The good thing here is that l3build handles almost everything with no effort at all.

Notes
-----

As the existing build system using log2tlg does not remove blank lines, a couple of
the tests don't pass with l3build unless you either locally modify l3build or the
.tlg files. Once that is done, everything should go to completion (i.e.
this can now do a release).

Text files are inclued in the result if you have a bang-up-to-date l3build
in your local tree (I added .txt as a standard extension today). This will
include all of the text files in the CTAN upload as extensionless files,
which makes sense as they will have Unix line endings in any case.

A TDS-style zip is not included in the upload but doing it would only
require changing one switch in the config Lua file.