Add initial version of 'Unicode data script'
At present, this extracts only case (\lccode/\uccode) data from 
UnicodeData.txt and not the XeTeX-specific line breaking info: that will
be added in a separate commit. The extracted data file is also not checked
in (it will need to be assuming we don't want to require the raw data be
available to build a format), and thus there is currently no change to the
kernel source files. Once all of the data extration is done that step will
be undertaken.